Make SVN++ compile and work without warnings in maintainer mode on Linux/g++.
* build.conf (svnxx-tests): Repeat the library dependencies of libsvnxx,
   since our generator isn't smart enough to inherit its dependencies
   at link time.

[in subversion/bindings/cxx]
* include/svnxx/exception.hpp (message::message): Fix -Wshadow.
* src/aprwrap/pool.hpp (pool::safe_alloc): Likewise.
* src/exception.cpp
  (checked_call..error_deleter): Likewise.
  (handle_one_error..message_builder::message_builder): Likewise.

* tests/test_revision.cpp (assignment): Do not rely on the system_clock's
   time_point precision being smaller than the time we store in svn::revision.
   Cast it to the nearest second instead.

* tests/test_strings.cpp (generate_string_data): Fix -Wsign-compare.



git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1850398 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/bindings/cxx/tests/test_strings.cpp
@@ -142,11 +142,12 @@ inline std::vector<std::basic_string<C>> generate_string_data(int count)
       std::u32string val;
       val.reserve(len);
 
-      for (int j = 0; j < len; ++j)
+      for (unsigned j = 0; j < len; ++j)
         {
         repeat:
           auto c = cgen(mt);
-          if (c >= cp::surrogate_min && c <= cp::surrogate_max)
+          if (uint_least32_t(c) >= cp::surrogate_min
+              && uint_least32_t(c) <= cp::surrogate_max)
             goto repeat;
           val.push_back(c);
         }
